API/requirements.txt
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
langchain
python-dotenv
openai
beautifulsoup4
numpy
tiktoken
faiss-cpu
uvicorn
fastapi
slack-sdk
slackeventsapi
slack-bolt
supabase
gotrue
httpx
websockets

